Breaking up the functionality into separate files.
