+++ py_pushover/Constants/Sounds.py
@@ -5,24 +5,24 @@ This is to be used with the 'sounds' argument of the PushOverManager.push_notifi
 
 see also: https://pushover.net/api#sounds
 """
-Short_Pushover = 'pushover'
-Short_Bike = 'bike'
-Short_Bugle = 'bugle'
-Short_Cash_Register = 'cashregister'
-Short_Classical = 'classical'
-Short_Cosmic = 'cosmic'
-Short_Gamelan = 'gamelan'
-Short_Incoming = 'incoming'
-Short_Intermission = 'intermission'
-Short_Magic = 'magic'
-Short_Mechanical = 'mechanical'
-Short_Piano_Bar = 'pianobar'
-Short_Siren = 'siren'
-Short_Space_Alarm = 'spacealarm'
-Short_Tug_Boat = 'tugboat'
-Long_Alien_Alarm = 'alien'
-Long_Climb = 'climb'
-Long_Persistent = 'persistent'
-Long_Pushover_Echo = 'echo'
-Long_Up_Down = 'updown'
-Silent = 'none'
+SHORT_PUSHOVER = 'pushover'
+SHORT_BIKE = 'bike'
+SHORT_BUGLE = 'bugle'
+SHORT_CASH_REGISTER = 'cashregister'
+SHORT_CLASSICAL = 'classical'
+SHORT_COSMIC = 'cosmic'
+SHORT_GAMELAN = 'gamelan'
+SHORT_INCOMING = 'incoming'
+SHORT_INTERMISSION = 'intermission'
+SHORT_MAGIC = 'magic'
+SHORT_MECHANICAL = 'mechanical'
+SHORT_PIANO_BAR = 'pianobar'
+SHORT_SIREN = 'siren'
+SHORT_SPACE_ALARM = 'spacealarm'
+SHORT_TUG_BOAT = 'tugboat'
+LONG_ALIEN_ALARM = 'alien'
+LONG_CLIMB = 'climb'
+LONG_PERSISTENT = 'persistent'
+LONG_PUSHOVER_ECHO = 'echo'
+LONG_UP_DOWN = 'updown'
+SILENT = 'none'
